DMP'25/AWS S3/requirements.txt
flask
flask-cors
psycopg2-binary
boto3
PyJWT
python-dotenv